✏️ WV-25 rename : 케밥케이스로 수정정
